Implemented basic ArnoldC variable operations
ArnoldC variables are instances of the class ArnoldCVariable which
stores a variable name and a value. These ArnoldC variables are kept
in the innermost function's closure.

Apart from variables, ArnoldC functions and statements, which are
respectively instances of the classes ArnoldCFunction and ArnoldCStatement,
have been partlially implemented. As of now, only the main function can
be defined. Furthermore, the only statements that are currently implemented
are those which define and set a variable's initial value, as well as
a statement that prints to stdout.