Added evaluation of conditional statements
ArnoldC conditionals are represented as instances of the class
ArnoldCConditional. On the one hand, the former is simillar to
ArnoldCStatement as it is pushed into the ArnoldC function's body
and later evaluated. On the other hand, conditionals are comparable to
ArnoldC functions as they also contain a body where statements are pushed
to instead of the innermost function.