Interpreter can define custom void and non-void functions with parameters
ArnoldC functions' parameter names are initially store in an array so
that their order could be preserved. When the function is called
parameter names are associated with the corresponding values passed to
that function call. Furthermore, the associated parameter name - value pair
is stored as a Ruby hash which is searched before the function's closure
when evaluating an ArnoldC expression (e.g. getting a variable's value).

Functions can use return statements. The latter ends the execution of
the function body. If the function is non-void i.e. returns a value, that
value is assigned to a variable declared before the function call.