Improved visibility by abstracting some repeating code
Most ArnoldC expressions are considered ArnoldC statements by the
Interpreter. Each of the mentioned methods create a new ArnoldCStatement,
define the former's body called when the statement is executed, and finally
push the instance to the innermost function. The aforementioned operations
are now defined outside the body of the Ruby method which represents an
ArnoldC statement via the use of metaprogramming (see ArnoldCPM::Interpreter's
singleton method 'extended').