Redefined the way the Interpreter defines functions
There is now one outermost function declared under the name '__program__'.
It contains instances of function templates and functions.

- Function templates
Function templates serve to preemptively declare a function and what said
function accoplishes. Templates contain only partial information of the
full function definition such as: name, parameter names (but not any
concrete values associated with those names), whether the function returns
a value or not, and the function body (i.e. the collection of linearly
evaluated ArnoldC statements). When function templates are defined
statements which create functions from said templates are also appended
to the innermost defined function.

- Functions
Functions are now instantiated via their corresponding function template.
Consequently, every time a function defines an inner function it would
be a different Ruby object and therefore a different ArnoldC object.

- General
Statements and conditionals no longer keep a local copy of the current
scope they are instantiated in. Instead both now contain a reference to
the function template they are included in. As a consequence the
Interpreter no longer depends on the scope copy that statements and
conditionals used to keep but use the template to achieve the same
functionallity in a much more simplistic way.
Statements and Conditionals have been aggusted to operate under this new
method of defining ArnoldC functions. Their functionallity has not been
hindered.