(Slightly) redefined the way the Interpreter defines functions (again)
- function stack
The newly-introduced function stack piles on top of one another instances
of ArnoldCFunction. As such many of these instances can have the same
function template. The function stack was made mainly so that proper
recurssion behaivour can be exhibited by the ArnoldC Interpreter as
it was difficult to keep track of and store all return values in a
singular instance of ArnoldCFunction.

- creation and invocation of functions
While reading the ArnoldC code the Interpreter creates only function
templates but no functions which can be executed. After that, the main
function is instantiated, pushed to the function stack and executed.
ArnoldC functions are created during the program's execution from
templates which share the same name only when said functions are invoked
(i.e. in the do_it_now statement).
Consequently, functions can no longer accept as parameters or return
instances of ArnoldCFunction. Instead the former can now receive and
return function templates. As a result, functions do not redefine their
inner functions anymore.

- logical operations
Logical 'or' now returns the first operand whose value evaluates to truth.
Similarly, logical 'and' returns the second operand whose value evaluates
to untruth.

- general
Altered the Ruby code to fit the changes to the Interpreter's behaviour.